rbml: move the tagged encoder/decoder to librustc_metadata.
